-Separated classes by files -Plugins folder with dlls
