+++ Program.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Configuration;
 using System.IO;
+using System.Reflection;
 using Newtonsoft.Json;
+namespace Reflection;
 
 
 [AttributeUsage(AttributeTargets.Property)]
@@ -18,130 +20,51 @@ public class ConfigurationItemAttribute : Attribute
 }
 
 
-public interface IConfigurationProvider
-{
-    void SaveSetting(string key, string value);
-    string LoadSetting(string key);
-}
-
-
-public class FileConfigurationProvider : IConfigurationProvider
-{
-    private readonly string filePath;
-
-    public FileConfigurationProvider(string filePath)
-    {
-        this.filePath = filePath;
-    }
-
-    public void SaveSetting(string key, string value)
-    {
-        var settings = File.Exists(filePath) ? JsonConvert.DeserializeObject<Dictionary<string, string>>(File.ReadAllText(filePath)) : new Dictionary<string, string>();
-        settings[key] = value;
-        File.WriteAllText(filePath, JsonConvert.SerializeObject(settings));
-    }
-
-
-    public string LoadSetting(string key)
-    {
-        if (File.Exists(filePath))
-        {
-            var settings = JsonConvert.DeserializeObject<Dictionary<string, string>>(File.ReadAllText(filePath));
-            if (settings.TryGetValue(key, out var value))
-            {
-                return value;
-            }
-        }
-        return null;
-    }
-}
-
-public class ConfigurationManagerConfigurationProvider : IConfigurationProvider
-{
-    private readonly Configuration configuration;
-
-    public ConfigurationManagerConfigurationProvider()
-    {
-        configuration = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.None);
-    }
-    public void SaveSetting(string key, string value)
-    {
-        if (configuration.AppSettings.Settings[key] != null)
-        {
-            configuration.AppSettings.Settings[key].Value = value;
-        }
-        else
-        {
-            configuration.AppSettings.Settings.Add(key, value);
-        }
-
-        configuration.Save(ConfigurationSaveMode.Modified);
-        ConfigurationManager.RefreshSection("appSettings");
-    }
-
-    public string LoadSetting(string key)
-    {
-        return configuration.AppSettings.Settings[key]?.Value;
-    }
-}
-
-public class ConfigurationComponentBase
-{
-    private readonly IConfigurationProvider provider;
-
-    public ConfigurationComponentBase(IConfigurationProvider provider)
-    {
-        this.provider = provider;
-    }
-
-    protected T GetSetting<T>(string key)
-    {
-        var value = provider.LoadSetting(key);
-        return (value != null) ? (T)Convert.ChangeType(value, typeof(T)) : default(T);
-    }
-
-    protected void SaveSetting(string key, string value)
-    {
-        provider.SaveSetting(key, value);
-    }
-}
-
-
-public class AppConfig : ConfigurationComponentBase
-{
-    [ConfigurationItemAttribute("LogLevel", "File")]
-    public int LogLevel
-    {
-        get => GetSetting<int>("LogLevel");
-        set => SaveSetting("LogLevel", value.ToString());
-    }
-
-    [ConfigurationItemAttribute("Timeout", "ConfigurationManager")]
-    public TimeSpan Timeout
-    {
-        get => TimeSpan.FromMilliseconds(GetSetting<double>("Timeout"));
-        set => SaveSetting("Timeout", value.TotalMilliseconds.ToString());
-    }
-
-    public AppConfig(IConfigurationProvider provider) : base(provider)
-    {
-    }
-}
-
 class Program
 {
+    static IConfigurationProvider[] LoadFromPlugin()
+    {
+        var plugins = new List<IConfigurationProvider>();
+        foreach (var file in Directory.GetFiles(@".\Plugins", "*.dll"))
+        {
+            try
+            {
+                var assembly = Assembly.LoadFrom(Directory.GetCurrentDirectory() + file);
+
+                foreach (var type in assembly.GetTypes())
+                {
+                    if (type.GetInterfaces().Contains(typeof(IConfigurationProvider)))
+                    {
+                        var plug = Activator.CreateInstance(type) as IConfigurationProvider;
+                        plugins.Add(plug);
+                    }
+                }
+
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine("Error while loading plugin");
+            }
+        }
+
+
+        return plugins.ToArray();
+
+    }
     static void Main()
     {
-        var fileProvider = new FileConfigurationProvider("/config.json");
-        var appConfigFile = new AppConfig(fileProvider);
+        var fileProvider = LoadFromPlugin();
 
-        appConfigFile.LogLevel = 3;
-        Console.WriteLine($"Log Level: {appConfigFile.LogLevel}");
+        foreach (var provider in fileProvider)
+        {
+            Console.WriteLine(provider);
+            var appConfigFile = new AppConfig(provider);
 
-        var configManagerProvider = new ConfigurationManagerConfigurationProvider();
-        var appConfigManager = new AppConfig(configManagerProvider);
+            appConfigFile.LogLevel = 3;
+            appConfigFile.Timeout = TimeSpan.FromSeconds(10);
+            Console.WriteLine($"Log Level: {appConfigFile.LogLevel}");
+            Console.WriteLine($"Timeout: {appConfigFile.Timeout.TotalSeconds} seconds");
+        }
 
-        appConfigManager.Timeout = TimeSpan.FromSeconds(10);
-        Console.WriteLine($"Timeout: {appConfigManager.Timeout.TotalSeconds} seconds");
     }
 }
